feat: experiment with using borrower for euler
please do not review this deeply as its just a branch off experiment to ensure we have a picture of what a potential borrower implementation would look like.

overall doesn't seem like a great approach becuase:
* determining the "borrower" of the funds and/or making sure it is securely set is a major challenge.
triggerFlashLoan only gives us lender, token, and amount. token can probably considered MUST be the the actual token being borrowed (due to the design of flashloans/approve borrower function). There may be some wiggle room here as technically the Borrower code can be modified for euler, but its unknow nwhat off-chain impacts this could have. amount is needed whether we are doing Aave or Euler. So that leaves lender, which is intended to be the contract that issues the loan, and we would really like this to be the actual EVault contract. If we use "lender" as the user/beneficiary of the trade instead, then we have to have some way to lookup the EVault contract associated with a particular token (and because of how Euler's community structure works, their may be many). The other options to solve this are
use a prehook
* on repay this data will have to be signed by the user because th erepay beneficiary could be malliable by the solver
use extra calldata from callBackData, the final argument
* this would require for the euler code to extract the data from the field and return th ecallback with that data removed. I am almos tcertain it would require solver/driver modifications to make this work properly
* im pretty sure this has the same issue as the prehook method, the data must be signed when repaying, if the borrower is somehow allowed to be set incorrectly
certain operations of the borrow/repay functions require authenticated calls to execute. for example, to borrow, you need to call the authenticated enableCollateral and enableController on the EVC to permit creation of the loan. in our current testing implementation, there is a temporary authorization granted by IEVC.permit https://github.com/cowprotocol/euler-integration-contracts/blob/feat/initial/test/CowEvcWrapperTest.t.sol#L190 , but its unclear how this permit could be easily replicated with flashloan router. The EVC does support another function, setOperator https://evc.wtf/docs/contracts/technical-reference/contract.EthereumVaultConnector#setaccountoperator, which would allow for the user to specify in advance . But then that gives the EVC contract basically unchecked governing authority over the user's loan, funds, and collateral transfers, which would likely be a security issue (not an issues for the permit because the permitted EVC calls are signed and only for the one purpose the user calls for)
* the collateral that the user wants to borrow will need to be enabled prior to lending by calling enableCollateral, since the flashloan wrapper doesn't really know what the borrowed collateral is from its params.
* this is less of a "problem" per se, but the flashloan router was clearly not designed with Euler's "non flashloan" flow in mind. For example, on repay its going to be a "flashloan" call where no loan is actually taken, the funds to repay are sent from the settlement contract to the flashloan router (most likely), and then the flashloan router calls "repay" with the "amount" to flashloan. even though we dont see it now, we are probably going to see more issues like the above appear just because we are using th eflashloan router for something it fundamentally doesn't do...