chore: remove CowEvcWrapper and fix compile notes in remaining legacy (#16)
## Description

Remove the CowEvcWrapper and fix any compile notes that have yet to be
fixed.

## Context

We no longer use the CowEvcWrapper (in favor of the 3 specialized
wrappers), so lets get those removed.

In addition, there are a number of places where compilation
warnings/notes are being emitted, so we can get those fixed now and make
it policy to not have any notes in the CI.

Finally, we also remove dependencies on the original Cow repo as the
`CowWrapper` provides all the datastructures needed anyway and we should
only be using one set of data structures for compiler consistency
reasons.

## Out of Scope
there may be some slightly unnecessary changes in the PR in order to
pave the way for the wrappers. please let me know if this is concern and
I can recategorize.

## Testing Instructions
verify that `forge build --deny notes` succeeds (no warnings)