more changes
since the token to be returned is a EVC token, we can transfer by way of the EVC to access the funds. we also dont need helperRepay, so we can remove that as well. the result is one less approval that needs to be sent, less separate function surface area, and basically the same outcome in terms of trading

probably need to apply correspondingly to the collateral swap wrapper