fixes for close position wrapper security (#19)
## Description
Fixes for close position wrappers security.

## Context

This is just an intermediate PR to try and fix some issues identified
with the overall design of the close position wrapper. The design
choices here will be applied to the other wrappers as well.

* added `Inbox` as a way to receive assets coming out of the setltement
contract. Implementors will need to make sure their `receiver` is set
appropriately, toherwise the wrapper will fail to work as
intended/revert
* remove the `helperRepay` function
* Change what was previously the helperRepay function into a signed
permit that pulls funds from the user's subaccount
* move the subaccount match check to the beginning since its so
fundamental

## Out of Scope
This is just an immediate fix for general knowledge of the overhaul going on