implement alternate EIP1271 flow to regenerate the expected cow settlement
the big benefit of this flow is that it works with one less signature from the user, since it generates the needed signature from .
Its also mostly not too complex because all the needed fields (save for appData) were already included in the verification.

One possible downside of this method is that only one order (at least in this specific impl) could be submitted for the user for one
account at a time.

another huge downside of this flow is that (at least in current implementation) its not required for the user's original address to be derivable from the signtaure field.